Seven Bag is written. Somewhat tested, testbench has been added
Some errors can be seen when removing pieces from the queue. Unsure if this is
due to lack of synchronization or if it's errors from the LFSRs. Suspect it is
the former but it is worth testing more from a gamer perspective.

Next priority is getting game states working. Game needs start/stop with
associated screens or effects.