Optimized floor check in GameStatesFSM
Given that the ghost tile and the falling tile are the same type of tile and
necessarily in the same orientation, it is unnecessary and therefore wasteful to
compare the tiles that they occupy and check that they all match. Instead, it is
more efficient to compare the coordinates instead. Untested, but saves ~800 LEs